establishing environment.
this commit was to focus on the establishment of the A2 environment. Removed all HTML references and started writing in PHP for this task.
Likewise, this was an early build without many of the PHP features been implemented. Next commit will contain updates relating to the PHP and establishment of the SQL server.